// src/store/types/mortgageCalc/enums.ts
export enum TypeEnum {
  BY_LOAN_AMOUNT = 'by-loan-amount',
  BY_PROPERTY_VALUE = 'by-property-value',
}

export enum ActionEnum {
  SET_CREDIT_AMOUNT = 'SET_CREDIT_AMOUNT',
  SET_CREDIT_TERM = 'SET_CREDIT_TERM',
  SET_INITIAL_FEE = 'SET_INITIAL_FEE',
  SET_PAYMENT_TYPE = 'SET_PAYMENT_TYPE',
  SET_PRICE = 'SET_PRICE',
  SET_RATE = 'SET_RATE',
  SET_TYPE = 'SET_TYPE',
}